Create optimal_steering.py
The optimal heading angle can be calculated by the angle of third waypoint (our destination) from the current coordinates, and then the cosine of heading difference is given as reward for alignment which will range from (-1 to 1, maximum at cos(0) = 1) which will be further finetuned with the weighted sum of all the rewards at return statement. This way you can maximize the reward for optimal heading direction of steering according to the dynamic path you just calculated for the future steps.